backend/requirements.txt
aniso8601==9.0.0
bcrypt==3.2.0
blinker==1.4
cffi==1.14.5
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.0.2
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-RESTful==0.3.8
Flask-WTF==0.14.3
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mongoengine==0.22.1
pycparser==2.20
PyJWT==2.0.1
pymongo==3.11.3
pytz==2021.1
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.3
